### 2025-10-27 #9311 @editorialbot
[PRE REVIEW]: Interfere: Studying Intervention Response Prediction in Complex Dynamic Models
**Submitting author:** @djpasseyjr (<a href="http://orcid.org/0000-0002-9310-3361">David Passey</a>)
**Repository:** https://github.com/djpasseyjr/interfere.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
