### 2025-11-03 d2b66b9
Rework summary.
