+++ paper/paper.md
@@ -42,7 +42,7 @@ bibliography: paper.bib
 ---
 # Summary
 
-The vision of Interfere is simple: What if we used high-quality scientific models to create causal dynamic benchmark scenarios? Randomized experimental data and intervention response time series are essential for testing methods that attempt to infer dynamic relationships from data, but obtaining such datasets can be expensive and difficult. Mechanistic models are commonly developed to simulate scenarios and predict the response of systems to interventions across economics, neuroscience, ecology, systems biology and other areas [@brayton_frbus_2014; @izhikevich_large-scale_2008; @banks_parameter_2017; @baker_mechanistic_2018]. Because these models are calibrated to the real world, they have the ability to generate diverse, complex, synthetic intervention responses that are characteristic of the real processes they emulate. Interfere offers the first steps towards this vision by combining (1) a general interface for simulating the effect of interventions on dynamic models, (2) a suite of predictive methods and cross validated hyper parameter optimization tools, and (3) the first known [extensible benchmark data set](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) of dynamic intervention response scenarios see Figure \ref{fig:sixty_models}.
+The Interfere package is designed to research intervention response prediction, that is, predicting how a complex, dynamic system will respond to a never-before-seen intervention. A hurdle to constructing methods to solve this problem is that the ideal data for benchmarking such methods comes from controlled experiment which can be prohibitively expensive, harmful or otherwise infeasible. As an alternative, we highlight that many disciplines, economics, neuroscience, ecology, systems biology and others, employ mechanistic dynamic models as an integral part of their research and application toolkit [@brayton_frbus_2014; @izhikevich_large-scale_2008; @banks_parameter_2017; @baker_mechanistic_2018]. In light of this, we propose the following: Insofar as a mechanistic model captures real characteristics of the process it emulates, the accuracy whereby a predictive algorithm is able to forecast the intervention response of that mechanistic model correlates with the predictive algorithm's ability to forecast the intervention response for similar real world systems. Interfere offers the first steps towards this vision by combining (1) a general interface for simulating the effect of interventions on dynamic models, (2) a suite of predictive methods and cross validated hyper parameter optimization tools, and (3) the first known [extensible benchmark data set](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) of dynamic intervention response scenarios see Figure \ref{fig:sixty_models}.
 
 ![Three-dimensional trajectories of sixty scenarios simulated with the Interfere
 package. The models simulated here are either differential equations or discrete time
