### 2025-11-10 05b28de
Rewrites to paper intro.
