+++ paper/paper.md
@@ -42,7 +42,9 @@ bibliography: paper.bib
 ---
 # Summary
 
-The Interfere package is designed to research intervention response prediction, that is, predicting how a complex, dynamic system will respond to a never-before-seen intervention. A hurdle to constructing methods to solve this problem is that the ideal data for benchmarking such methods comes from controlled experiment which can be prohibitively expensive, harmful or otherwise infeasible. As an alternative, we highlight that many disciplines, economics, neuroscience, ecology, systems biology and others, employ mechanistic dynamic models as an integral part of their research and application toolkit [@brayton_frbus_2014; @izhikevich_large-scale_2008; @banks_parameter_2017; @baker_mechanistic_2018]. In light of this, we propose the following: Insofar as a mechanistic model captures real characteristics of the process it emulates, the accuracy whereby a predictive algorithm is able to forecast the intervention response of that mechanistic model correlates with the predictive algorithm's ability to forecast the intervention response for similar real world systems. Interfere offers the first steps towards this vision by combining (1) a general interface for simulating the effect of interventions on dynamic models, (2) a suite of predictive methods and cross validated hyper parameter optimization tools, and (3) the first known [extensible benchmark data set](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) of dynamic intervention response scenarios see Figure \ref{fig:sixty_models}.
+The Interfere package is designed to research *intervention response prediction*, or, more specifically, predicting how a complex, dynamic system will respond to a never-before-seen intervention. When developing predictive methods capable of solving this problem, ideal benchmarking data comes exclusively from carefully controlled, longitudinal experiments. Unfortunately, unavoidable confounding, prohibitive costs and ethical boundaries often stymie attempts to run controlled experiments for many important problems. Synthetic data can be used as an alternative to experimental data and has been employed to benchmark predictive algorithms [cite causal graph etc]. However, in practice, the synthetic data used for benchmarking is often simplistic, and the simulation models used to generate it show a range of familiar shortcomings: static, linear relationships, overly independent noise profiles, and a lack of the complex mechanistic relationships and feedback loops that we find in the real world. As a solution, we propose a middle ground: Scientists in many disciplines, such as economics, ecology, systems biology and others, employ mechanistic dynamic models as an integral part of their research and application toolkit [@brayton_frbus_2014; @izhikevich_large-scale_2008; @banks_parameter_2017; @baker_mechanistic_2018]. Many of these models are accompanied by a body of evidence demonstrating that they do indeed capture important characteristics of the real world problems they emulate. Instead of using simple synthetic data for benchmarking, we propose using datasets generated by scientific models of complex systems as a ground truth against which we can benchmark and test methods that predict intervention response. The logic is this: if a scientific model captures important characteristics of the world and a method can accurately predict intervention response *for that model* then the method ought to be able to predict intervention response for similar systems in the real world.
+
+Interfere offers the first steps towards this vision by combining (1) a general interface for simulating the effect of interventions on dynamic models, (2) a suite of predictive methods and cross validated hyper parameter optimization tools, and (3) the first known [extensible benchmark data set](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) of dynamic intervention response scenarios see Figure \ref{fig:sixty_models}.
 
 ![Three-dimensional trajectories of sixty scenarios simulated with the Interfere
 package. The models simulated here are either differential equations or discrete time
