+++ paper/paper.bib
@@ -43,6 +43,7 @@
     author = {Brayton, Flint and Laubach, Thomas and Reifschneider, David},
     year = {2014},
     pages = {03},
+	doi={10.17016/2380-7172.0012},
 }
 
 @book{imbens_causal_2015,
@@ -191,6 +192,7 @@
 	year = {1995},
 	note = {Publisher: Elsevier},
 	pages = {217--230},
+	doi = {10.1006/jtbi.1995.0058},
 }
 
 @incollection{kaneko_coupled_1991,
@@ -211,6 +213,7 @@
 	year = {2022},
 	note = {Publisher: Wiley Online Library},
 	pages = {6086--6098},
+	doi = {10.1111/febs.16124},
 }
 
 @article{hening_stochastic_2018,
@@ -260,6 +263,7 @@
 	publisher = {Springer Berlin Heidelberg},
 	author = {Gardiner, Crispin},
 	year = {2009},
+	doi = {10.1093/hesc/9780192896599.003.0008},
 }
 
 @article{belozyorov_exponential_2015,
@@ -271,6 +275,7 @@
 	year = {2015},
 	note = {Publisher: World Scientific},
 	pages = {1550048},
+	doi = {10.1142/S0218127415500480},
 }
 
 @article{liping_new_2021,
@@ -282,6 +287,7 @@
 	year = {2021},
 	note = {Publisher: Elsevier},
 	pages = {5193--5204},
+	doi = {10.1016/j.aej.2021.04.023},
 }
 
 @incollection{lorenz_deterministic_2017,
@@ -291,6 +297,7 @@
 	author = {Lorenz, Edward N},
 	year = {2017},
 	pages = {367--378},
+	doi = {10.1201/9780203734636-38},
 }
 
 @article{rossler_equation_1976,
@@ -313,6 +320,7 @@
 	year = {1999},
 	note = {Publisher: World Scientific},
 	pages = {1889--1905},
+	doi = {10.1142/S0218127499001383},
 }
 
 @book{hamilton_time_2020,
@@ -331,6 +339,7 @@
 	year = {1972},
 	note = {Publisher: Elsevier},
 	pages = {1--24},
+	doi = {10.1016/S0006-3495(72)86068-5},
 }
 
 @article{black_pricing_1973,
@@ -353,6 +362,7 @@
 	year = {2006},
 	note = {Publisher: Elsevier},
 	pages = {1370--1383},
+	doi = {10.1016/j.amc.2005.05.045},
 }
 
 @misc{garza2022statsforecast,
@@ -393,6 +403,7 @@ journal = {Journal of Open Source Software}
   author={Seabold, Skipper and Perktold, Josef},
   booktitle={9th Python in Science Conference},
   year={2010},
+  doi = {10.25080/majora-92bf1922-011},
 }
 
 @inproceedings{akiba2019optuna,
