### 2025-11-17 b57ffad
Cut down text.
