+++ paper/paper.md
@@ -47,19 +47,19 @@ The Interfere package is designed to research *intervention response prediction*
 Interfere offers the first steps towards this vision by combining (1) a general interface for simulating the effect of interventions on dynamic models, (2) a suite of predictive methods and cross validated hyper parameter optimization tools, and (3) the first known [extensible benchmark data set](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) of dynamic intervention response scenarios see Figure \ref{fig:sixty_models}.
 
 ![Three-dimensional trajectories of sixty scenarios simulated with the Interfere
-package. The models simulated here are either differential equations or discrete time
-difference equations. For each system, the trajectory in blue represents the natural behavior of
-the system and the red depicts how the system responds to a specified intervention.
-Many of the models pictured have more than three dimensions (in such cases,
-only the three dimensions of the trajectory with the highest variance
-are shown). These sixty scenarios make up the [Interfere Benchmark 1.1.1](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) for
-intervention response prediction which is available online for download.
+package. Simulated models are either differential equations or discrete time
+difference equations. Trajectories in blue represent the natural behavior of
+the system, while red depicts response to a specified intervention.
+For models with more than three dimensions,
+only the three dimensions with highest variance
+are shown. These sixty scenarios, making up the [Interfere Benchmark 1.1.1](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) for
+intervention response prediction, are available for download.
 \label{fig:sixty_models}](../images/sixty_models.png)
 
 # Statement of Need
 
 ![**Original System Trajectory (Left):** The natural, uninterupted evolution of
-the quaratic Belozyorov system [@belozyorov_exponential_2015] simulated using
+the quadratic Belozyorov system [@belozyorov_exponential_2015] simulated using
 the Interfere package with a small
 amount of stochastic noise.
 **System Trajectory After Intervention (Center):** The
@@ -77,62 +77,48 @@ intervention and makes an attempt to predict the intervention response (red
 curve).
 ](../images/interfere_usage_combined.png)
 
-Over the past twenty years, the scientific community has experienced the emergence
-of multiple frameworks for identifying causal relationships in observational
+Over the past twenty years, multiple frameworks have emerged for identifying causal relationships in observational
 data [@imbens_causal_2015; @pearl_causality_2009; @wieczorek_information_2019].
 The most influential frameworks are probabilistic and, while it is not a necessary
-condition for identifying causality, historically a static, linear relationship has
+condition for identifying causality, a static, linear relationship has
 often been assumed. However, when attempting to anticipate the response of complex dynamic
 systems in the medium and long term, a linear approximation of the dynamics can be
-insufficient.  Therefore, researchers have increasingly begun to employ
-non-linear, dynamic techniques for causal discovery and forecasting [e.g. @runge_discovering_2022]. Still,
+insufficient.  Therefore, 
+non-linear, dynamic techniques have been employed for causal discovery and forecasting [e.g. @runge_discovering_2022]. Nevertheless,
 there are relatively few techniques that are able to fit causal dynamic
-nonlinear models to data. Because of this, we see an opportunity to bring
-together the insights from recent advancements in causal inference with
-historical work in dynamic modeling and simulation.
+nonlinear models to data. 
 
[…]
