### 2025-11-17 e73ae2a
Cut down text of intro.
