+++ paper/paper.md
@@ -42,7 +42,31 @@ bibliography: paper.bib
 ---
 # Summary
 
-The Interfere package is designed to research *intervention response prediction*, or, more specifically, predicting how a complex, dynamic system will respond to a never-before-seen intervention. When developing predictive methods capable of solving this problem, ideal benchmarking data comes exclusively from carefully controlled, longitudinal experiments. Unfortunately, unavoidable confounding, prohibitive costs and ethical boundaries often stymie attempts to run controlled experiments for many important problems. Synthetic data can be used as an alternative to experimental data and has been employed to benchmark predictive algorithms [cite causal graph etc]. However, in practice, the synthetic data used for benchmarking is often simplistic, and the simulation models used to generate it show a range of familiar shortcomings: static, linear relationships, overly independent noise profiles, and a lack of the complex mechanistic relationships and feedback loops that we find in the real world. As a solution, we propose a middle ground: Scientists in many disciplines, such as economics, ecology, systems biology and others, employ mechanistic dynamic models as an integral part of their research and application toolkit [@brayton_frbus_2014; @izhikevich_large-scale_2008; @banks_parameter_2017; @baker_mechanistic_2018]. Many of these models are accompanied by a body of evidence demonstrating that they do indeed capture important characteristics of the real world problems they emulate. Instead of using simple synthetic data for benchmarking, we propose using datasets generated by scientific models of complex systems as a ground truth against which we can benchmark and test methods that predict intervention response. The logic is this: if a scientific model captures important characteristics of the world and a method can accurately predict intervention response *for that model* then the method ought to be able to predict intervention response for similar systems in the real world.
+The Interfere package focuses on *intervention response prediction*—forecasting how complex, dynamic systems respond to novel interventions. When developing predictive
+methods capable of solving this problem, ideal benchmarking data comes from
+controlled, longitudinal experiments. Such experiments are often infeasible due
+to confounding, cost, and ethical constraints. 
+
+Synthetic benchmarks, which are often used in response to this problem [cite], typically rely on models
+with static linear relationships and overly independent noise profiles that lack
+the complex interdependent feedback we find in the real world.
+
+As a solution, we propose a middle ground: Scientists in many disciplines, such
+as economics, ecology, systems biology and others, employ mechanistic dynamic
+models as a part of their research and application toolkit [@brayton_frbus_2014;
+@izhikevich_large-scale_2008; @banks_parameter_2017; @baker_mechanistic_2018].
+Many of these models are accompanied by a body of evidence demonstrating that
+they do indeed capture important characteristics of the real world problems they
+emulate. 
+Instead of using simple synthetic data for benchmarking, we propose using data
+generated by detailed scientific models as a ground truth against which we can
+benchmark methods that predict intervention response. 
+
+The logic is this: Methods that accurately predict intervention responses of
+scientific models should generalize to real-world processes insofar as these
+models capture them.
+
+
 
 Interfere offers the first steps towards this vision by combining (1) a general interface for simulating the effect of interventions on dynamic models, (2) a suite of predictive methods and cross validated hyper parameter optimization tools, and (3) the first known [extensible benchmark data set](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) of dynamic intervention response scenarios see Figure \ref{fig:sixty_models}.
 
