+++ paper/paper.md
@@ -48,7 +48,7 @@ controlled, longitudinal experiments. Such experiments are often infeasible due
 to confounding, cost, and ethical constraints. 
 
 Synthetic benchmarks, which are often used in response to this problem [cite], typically rely on models
-with static linear relationships and overly independent noise profiles that lack
+with static linear relationships and/or overly independent noise profiles that lack
 the complex interdependent feedback we find in the real world.
 
 As a solution, we propose a middle ground: Scientists in many disciplines, such
@@ -57,16 +57,7 @@ models as a part of their research and application toolkit [@brayton_frbus_2014;
 @izhikevich_large-scale_2008; @banks_parameter_2017; @baker_mechanistic_2018].
 Many of these models are accompanied by a body of evidence demonstrating that
 they do indeed capture important characteristics of the real world problems they
-emulate. 
-Instead of using simple synthetic data for benchmarking, we propose using data
-generated by detailed scientific models as a ground truth against which we can
-benchmark methods that predict intervention response. 
-
-The logic is this: Methods that accurately predict intervention responses of
-scientific models should generalize to real-world processes insofar as these
-models capture them.
-
-
+emulate. We propose using datasets generated by such models of complex systems to benchmark methods that predict intervention response. If a scientific model captures important characteristics of the world and a method can accurately predict intervention response *for that model* then the method might also predict intervention response for similar real-world systems.
 
 Interfere offers the first steps towards this vision by combining (1) a general interface for simulating the effect of interventions on dynamic models, (2) a suite of predictive methods and cross validated hyper parameter optimization tools, and (3) the first known [extensible benchmark data set](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) of dynamic intervention response scenarios see Figure \ref{fig:sixty_models}.
 
@@ -76,8 +67,7 @@ difference equations. Trajectories in blue represent the natural behavior of
 the system, while red depicts response to a specified intervention.
 For models with more than three dimensions,
 only the three dimensions with highest variance
-are shown. These sixty scenarios, making up the [Interfere Benchmark 1.1.1](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) for
-intervention response prediction, are available for download.
+are shown. These sixty scenarios, make up the downloadable [Interfere Benchmark 1.1.1](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing).
 \label{fig:sixty_models}](../images/sixty_models.png)
 
 # Statement of Need
@@ -109,10 +99,17 @@ often been assumed. However, when attempting to anticipate the response of compl
 systems in the medium and long term, a linear approximation of the dynamics can be
 insufficient.  Therefore, 
 non-linear, dynamic techniques have been employed for causal discovery and forecasting [e.g. @runge_discovering_2022]. Nevertheless,
-there are relatively few techniques that are able to fit causal dynamic
-nonlinear models to data. 
+there are relatively few intervention oriented techniques that are able to fit
+dynamic nonlinear models to data. 
 
[…]
