### 2025-11-24 2d5aa22
Trim content of paper.
