+++ paper/paper.md
@@ -42,104 +42,68 @@ bibliography: paper.bib
 ---
 # Summary
 
-The Interfere package focuses on *intervention response prediction*—forecasting how complex, dynamic systems respond to novel interventions. When developing predictive
-methods capable of solving this problem, ideal benchmarking data comes from
-controlled, longitudinal experiments. Such experiments are often infeasible due
-to confounding, cost, and ethical constraints. 
+The Interfere package focuses on *intervention response prediction*—forecasting how complex, dynamic systems respond to novel interventions. 
+While ideal data for benchmarking predictive methods for such problems would come from controlled, longitudinal experiments, such experiments are often infeasible due to confounding, cost, and ethical constraints. 
+In contrast, many synthetic benchmarks rely on models with static linear relationships and/or independent noise profiles that lack the complex interdependent feedback present in the real world.
+However, many disciplines, including economics, ecology, and systems biology, employ mechanistic dynamic models [@brayton_frbus_2014; @izhikevich_large-scale_2008; @banks_parameter_2017; @baker_mechanistic_2018]
+with evidence demonstrating they capture important real-world characteristics. 
+We propose using datasets generated by such complex systems to benchmark methods that predict intervention response. 
+If a model captures important real-world characteristics and a method can accurately predict intervention response *for that model* then the method might also predict intervention response for similar real-world systems.
+Interfere works towards this vision by combining (1) a general interface for simulating interventions on dynamic models, (2) a suite of predictive methods and cross-validated hyperparameter optimization tools, and (3) an [extensible benchmark data set](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) of dynamic intervention response scenarios (see Figure \ref{fig:sixty_models}).
 
-Synthetic benchmarks, which are often used in response to this problem [cite], typically rely on models
-with static linear relationships and/or overly independent noise profiles that lack
-the complex interdependent feedback we find in the real world.
-
-As a solution, we propose a middle ground: Scientists in many disciplines, such
-as economics, ecology, systems biology and others, employ mechanistic dynamic
-models as a part of their research and application toolkit [@brayton_frbus_2014;
-@izhikevich_large-scale_2008; @banks_parameter_2017; @baker_mechanistic_2018].
-Many of these models are accompanied by a body of evidence demonstrating that
-they do indeed capture important characteristics of the real world problems they
-emulate. We propose using datasets generated by such models of complex systems to benchmark methods that predict intervention response. If a scientific model captures important characteristics of the world and a method can accurately predict intervention response *for that model* then the method might also predict intervention response for similar real-world systems.
-
-Interfere offers the first steps towards this vision by combining (1) a general interface for simulating the effect of interventions on dynamic models, (2) a suite of predictive methods and cross validated hyper parameter optimization tools, and (3) the first known [extensible benchmark data set](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) of dynamic intervention response scenarios see Figure \ref{fig:sixty_models}.
-
-![Three-dimensional trajectories of sixty scenarios simulated with the Interfere
-package. Simulated models are either differential equations or discrete time
-difference equations. Trajectories in blue represent the natural behavior of
-the system, while red depicts response to a specified intervention.
-For models with more than three dimensions,
-only the three dimensions with highest variance
[…]
