### 2025-11-24 #9311 @djpasseyjr
@danielskatz Ok, took us a bit, but we've got the paper below word limit. (I also sorted out the DOIs.) Let me know if I've missed anything.

Thanks,
DJ
