### 2025-11-24 ce42b38
Adds pydantic for nixtla deps.
