### 2025-11-24 c6b6e54
Adds pydantic to models dependency group.
