### 2025-12-17 #9311 @mstimberg
Apologies @danielskatz, this notification somehow slipped through. Happy to take this!
