### 2026-01-16 #9792 @editorialbot
[REVIEW]: Interfere: Studying Intervention Response Prediction in Complex Dynamic Models
**Submitting author:** @djpasseyjr (<a href="http://orcid.org/0000-0002-9310-3361">David Passey</a>)
**Repository:** https://github.com/djpasseyjr/interfere.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
