### 2026-01-16 #9792 @mstimberg
@djpasseyjr unfortunately we are in the uncomfortable position that while you submitted your paper in 2025, the review only started in 2026, which means that the new 2026 JOSS guidelines apply to your work. See this blog post for more details: https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future In particular, the new guidelines require a slightly different structure of the paper, with additional sections (see this [editorialbot comment](https://github.com/openjournals/joss-reviews/issues/9792#issuecomment-3760821856) and the updated documentation: https://joss.readthedocs.io/en/latest/paper.html). Please update your manuscript accordingly and let us know when you've done so (@Jacob-Stevens-Haas, @yunus-kausable, please be aware that this means that the paper is not actually yet in its final form).

Also, given that you seem to be the only contributor to the GitHub repository, but that this paper has three additional authors, could you quickly state the contributions of the other authors? Of course, contributions that do not consist in writing code are acceptable, see https://joss.readthedocs.io/en/latest/submitting.html#authorship.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/9792#issue-3822710414). Please try to make a first review pass within 2–4 weeks, or let us know if you already know that this will not be possible.

Thank you for your time, and happy reviewing!
