### 2026-01-16 djpasseyjr/interfere#3 @yunus-kausable
[JOSS Review] Document how to run tests
Hi @djpasseyjr,

this is Yunus, I'm one of the reviewers of your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/9792). Super interesting submission, thank you … looking forward to reviewing!

I've started to go through the repository and the paper and will be opening issues in this project for anything that I'd like to bring up; some of these may just be questions or suggestions, for others it may be good to discuss a bit more. Please don't just change things because it's part of the review. It's your project after all.
I hope this process also works for you?

Below, some first small (purely infrastructural) things from trying to run the tests locally.

Let me know if you have any questions, please. Cheers :)

---

## Specify custom `surd` and `pyclustering` dependencies

You can just add them to the `dev` options in `pyproject.toml` :

```toml

dev = [
    # ...
    "pydantic",
    "surd @ git+https://github.com/djpasseyjr/surd.git",
    "pyclustering @ git+https://github.com/djpasseyjr/pyclustering"
]
```

You already have very good error messages, but given that you are using the `pyproject.toml` anyway I think this would smooth things out. Or was there a particular reason not to do this?

## Add a section to the `README.md` on how to run the tests.

Doesn't need to be a lot, perhaps just:

```md
## Running Tests
1. Follow installation instructions for the cloned repository, see above.
2. Install test dependencies using `pip install -e ".[dev]"`
3. Run tests: `python -m pytest -v tests`
```

The tests all passed for me, yay 🎉

## Also test more recent Python versions automatically
Afais, the GitHub action only tests this with Python 3.9 - 3.11. My local tests ran without problems on 3.13.

Automatically testing on more recent versions would be great, actually, at least up to 3.13. Note that [Python 3.9 is EOL](https://devguide.python.org/versions/) by now …

You can also add the following to your `pyproject.toml` (under `[project]`) to communicate [classifiers to PyPI](https://pypi.org/classifiers/) that show up on the PyPI project page (and in a badge, if you want):

```toml
  classifiers = [
      "Programming Language :: Python :: 3",
      "Programming Language :: Python :: 3.9",
      "Programming Language :: Python :: 3.10",
      "Programming Language :: Python :: 3.11",
      "Programming Language :: Python :: 3.12",
      "Programming Language :: Python :: 3.13",
      "License :: OSI Approved :: MIT License",
  ]

  requires-python = ">=3.9"
```
