### 2026-01-16 #9792 @yunus-kausable
Thanks, @mstimberg and @danielskatz! And thanks @djpasseyjr for the submission.

> but we will have to update this review thread to point to the old docs, and we'll have to be sure to change the checklists as they are generated to go back to the old checklists.

Just to be clear: Is this something that I should do manually (from some pre-2026 review issue) or will you take care of this?
