### 2026-01-17 #9792 @danielskatz
@djpasseyjr - your checklist is now the old checklist

> Sounds good. I'll get started with the tasks on my end!
