### 2026-01-17 #9792 @danielskatz
👋 @mstimberg - please let @arfon know when the second reviewer has created their checklist so that it can also be updated to the older checklist that we will use for this pre-2026 submission.
