### 2026-01-23 #9792 @mstimberg
> 👋 [@mstimberg](https://github.com/mstimberg) - please let [@arfon](https://github.com/arfon) know when the second reviewer has created their checklist so that it can also be updated to the older checklist that we will use for this pre-2026 submission.

Hi @danielskatz. The second reviewer has created their checklist, but it looks as if it is already correct? For the scope, it links to the pre-2026 guidelines at https://joss.readthedocs.io/en/pre-2026-submission/submitting.html#substantial-scholarly-effort, and it doesn't mention the new required sections.
