### 2026-01-26 7292fef
Updates to deps so tests are automatically runnable.
