+++ .github/workflows/python-package.yml
@@ -16,7 +16,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
 
     steps:
     - uses: actions/checkout@v3
@@ -29,8 +29,6 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install pyclustering@git+https://github.com/djpasseyjr/pyclustering
-        pip install git+https://github.com/djpasseyjr/surd.git
         pip install ".[dev]"
         pip install pytest-cov
 
