+++ pyproject.toml
@@ -34,6 +34,9 @@ dev = [
     "pytest",
     "sktime",
     "tigramite",
+    "surd @ git+https://github.com/djpasseyjr/surd.git",
+    "pyclustering @ git+https://github.com/djpasseyjr/pyclustering",
+    "pydantic",
     "pytest-cov",
     "pytest",
     "pylint",
