### 2026-02-01 djpasseyjr/interfere#8 @Jacob-Stevens-Haas
[JOSS Review] Doc suggestions
The docs read great and it looks straightforwards to use all the provided classes/functions! BZ

1) Recommend moving the Optimization page after the Prediction page, since the former references the latter
2) It could be useful to completely document the API for subclassing `ForecastMethod` so that it can be used in the hyperparameter optimization, which I think also requires `_get_optuna_params()` (or is this set via `CrossValObjective(hyperparam_func=?)`.  E.g. if I've got my own dynamics-learning model and I want to advertise its hyperparameters so `interfere` can use  hyperparameter tuning.  Or I want to use a `SINDy` object but the default `interfere._methods.SINDy` hyperparameters don't make sense for my choices (e.g. some optimizers don't have a [`threshold` or `alpha`](https://github.com/djpasseyjr/interfere/blob/c6b6e54c7c064c4249b74b62e3b5af9e2ae90372/interfere/_methods/sindy.py#L176)).

On a separate note... I owe you a small apology.  I have been shrinking the pysindy API over the past few years, and I allowed a breaking change with no deprecation in 2.1.0.  You should limit the pysindy version to `<2.1.0` or `!=2.1.0,<3.0`.  `discrete_time` is gone in favor of a separate `DiscreteSINDy` class.  The rationale is to make all class combinations type-safe and mathematically consistent (`WeakPDELibrary` will be the next to get converted, then `SINDy-PI`, then probably `PDELibrary`, all to be replaced with separate classes).
