### 2026-02-01 djpasseyjr/interfere#8 @Jacob-Stevens-Haas
Docstring to `SignalIntervention` looks like you need to finish writing [this line](https://github.com/djpasseyjr/interfere/blob/c6b6e54c7c064c4249b74b62e3b5af9e2ae90372/interfere/interventions.py#L110)
