### 2026-02-01 djpasseyjr/interfere#8 @Jacob-Stevens-Haas
In the Simulation page, you mention that the `DynamicModel` class handles, among other things

> Noise Injection: Builds and injects system (stochastic) and measurement noise

It isn't clear how the user specifies noise, or if the user isn't able to specify it, what type of default noise and noise level is added to trajectories during integration.  In particular, in the paper and README.md, you point out that it's easier to learn a deterministic system with noisy shocks than one without.  While the research details might be out of scope for JOSS, it does suggest that an important point about the API is choosing the stochasticity added to dynamics.
