### 2026-02-02 2d341af
Adds black settings and build system.
