### 2026-02-02 cd8b639
Black reformatting
