+++ interfere/dynamics/generative_forecasters.py
@@ -1,5 +1,4 @@
-"""Dynamic model wrapper for predictive algorithms. 
-"""
+"""Dynamic model wrapper for predictive algorithms."""
 
 from typing import Callable, Optional, Union
 
@@ -16,7 +15,6 @@ from ..utils import copy_doc
 
 
 class GenerativeForecaster(DynamicModel):
-
 
     def __init__(
         self,
@@ -44,8 +42,9 @@ class GenerativeForecaster(DynamicModel):
         self.fitted_method = fitted_method
         if not self.fitted_method.is_fit:
             raise ValueError(
-                f"{type(self).__name__}.__init__ requires a fitted method.")
-        
+                f"{type(self).__name__}.__init__ requires a fitted method."
+            )
+
         self.timestep = self.fitted_method.timestep_of_fit
 
         if self.timestep is None:
@@ -64,20 +63,21 @@ class GenerativeForecaster(DynamicModel):
 
         super().__init__(dim, measurement_noise_std, sigma)
 
-
     @copy_doc(DynamicModel.simulate)
     def _simulate(
         self,
         t: np.ndarray,
         prior_states: np.ndarray,
         prior_t: Optional[np.ndarray] = None,
-        intervention: Optional[Callable[[np.ndarray, float], np.ndarray]]= None,
+        intervention: Optional[
+            Callable[[np.ndarray, float], np.ndarray]
+        ] = None,
         rng: np.random.mtrand.RandomState = DEFAULT_RANGE,
-        **kwargs
+        **kwargs,
     ) -> np.ndarray:
-        
+
         num_prior_obs, dim = prior_states.shape
-        
+
         if self.fitted_method.endog_dim_of_fit != dim:
             raise ValueError(
                 f"{type(self).__name__}.simulate() was passed prior states with"
@@ -89,38 +89,38 @@ class GenerativeForecaster(DynamicModel):
         for i in range(len(t) - 1):
 
             new_states = self.fitted_method.predict(
-                t[i:(i+2)],
+                t[i : (i + 2)],
                 prior_states,
                 prior_t=prior_t,
-                prediction_max=np.inf
+                prediction_max=np.inf,
             )
-            next_state = new_states[-1,:]
+            next_state = new_states[-1, :]
 
             # Optionally intervene in model.
             if intervention is not None:
-                next_state = intervention(next_state, t[i+1])
+                next_state = intervention(next_state, t[i + 1])
 
             # Add stochastic noise.
             next_state += self.sigma @ rng.normal(size=self.dim)
 
             prior_states = np.vstack([prior_states, next_state])
-            prior_t = np.hstack([prior_t, [t[i+1]]])
+            prior_t = np.hstack([prior_t, [t[i + 1]]])
 
-        sim_states =  prior_states[-len(t):]
+        sim_states = prior_states[-len(t) :]
 
         # Add measurement noise to all but initial condition.
         sim_states[1:, :] = self.add_measurement_noise(
-            sim_states[1:, :], rng=rng)
-        
+            sim_states[1:, :], rng=rng
+        )
+
         return sim_states
-        
 
 
 def generative_lorenz_VAR_forecaster(
[…]
