+++ interfere/dynamics/ornstein_uhlenbeck.py
@@ -12,7 +12,7 @@ class OrnsteinUhlenbeck(StochasticDifferentialEquation):
         theta: Optional[np.ndarray] = None,
         mu: Optional[np.ndarray] = None,
         sigma: Optional[np.ndarray] = None,
-        measurement_noise_std: Optional[np.ndarray] = None
+        measurement_noise_std: Optional[np.ndarray] = None,
     ):
         """Initializes n-dimensional Ornstein Uhlenbeck process.
 
@@ -28,18 +28,20 @@ class OrnsteinUhlenbeck(StochasticDifferentialEquation):
                 model. For example, if the dynamic model had two variables x1
                 and x2 and `measurement_noise_std = [1, 10]`, then
                 independent gaussian noise with standard deviation 1 and 10
-                will be added to x1 and x2 respectively at each point in time.  
+                will be added to x1 and x2 respectively at each point in time.
         """
         # Set dimension
         super().__init__(len(mu), measurement_noise_std, sigma)
 
         # Input validation
-        if any([
-            mu.shape[0] != theta.shape[0],
-            theta.shape[0] != theta.shape[1],
-            theta.shape[0] != self.sigma.shape[0],
-            self.sigma.shape[1] != mu.shape[0]
-        ]):
+        if any(
+            [
+                mu.shape[0] != theta.shape[0],
+                theta.shape[0] != theta.shape[1],
+                theta.shape[0] != self.sigma.shape[0],
+                self.sigma.shape[1] != mu.shape[0],
+            ]
+        ):
             raise ValueError(
                 "Parameters for OrnsteinUhlenback must have matching "
                 "dimensions. Argument shapes: "
@@ -54,6 +56,6 @@ class OrnsteinUhlenbeck(StochasticDifferentialEquation):
 
     def drift(self, x: np.ndarray, t: float):
         return self.theta @ (self.mu - x)
-    
+
     def noise(self, x: np.ndarray, t: float):
         return self.sigma

[…]
