### 2026-02-02 62f812a
Autoformatter (black) and flake8 compliance updates. Cosmetic only.
