+++ interfere/_methods/vector_autoregression.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, Optional
 
 import numpy as np
 from statsmodels.tsa.api import VAR as smVAR
