### 2026-02-02 1d3ace2
Updates developer instructions with linter, formatting and testing.
