### 2026-02-03 #9792 @Jacob-Stevens-Haas
@mstimberg I've completed a first pass on the paper; unchecked boxes reflect minor revisions posted in the github for the author to respond to.  This being my first JOSS review, please let me know if I'm off base on anything!
