### 2026-02-09 a3431d0
Adds mkdocs instructions for building docs.
