### 2026-02-09 3c9100f
Pin pysindy version bc of breaking changes.
