+++ pyproject.toml
@@ -12,7 +12,7 @@ dependencies = [
     "networkx",
     "numpy<2",
     "optuna",
-    "pysindy",
+    "pysindy <2.1.0",
     "scikit-learn",
     "sdeint",
     "statsmodels",
