### 2026-02-09 00407f3
Updates docs in line with JOSS review.
