+++ interfere/dynamics/__init__.py
@@ -1,5 +1,8 @@
-from .arithmetic_brownian_motion import ArithmeticBrownianMotion  # noqa: F401
-from .base import OrdinaryDifferentialEquation, StochasticDifferentialEquation  # noqa: F401
+"""Dynamics models for Interfere."""
+# flake8: noqa
+
+from .arithmetic_brownian_motion import ArithmeticBrownianMotion
+from .base import OrdinaryDifferentialEquation, StochasticDifferentialEquation
 from .coupled_map_lattice import (
     coupled_logistic_map,
     StochasticCoupledMapLattice,
@@ -12,22 +15,22 @@ from .coupled_map_lattice import (
     coupled_map_1dlattice_spatiotemp_intermit1,
     coupled_map_1dlattice_spatiotemp_intermit2,
     coupled_map_1dlattice_traveling_wave,
-)  # noqa: F401
+)
 from .decoupled_noise_dynamics import (
     StandardNormalNoise,
     StandardCauchyNoise,
     StandardExponentialNoise,
     StandardGammaNoise,
     StandardTNoise,
-)  # noqa: F401
-from .geometric_brownian_motion import GeometricBrownianMotion  # noqa: F401
-from .kuramoto import Kuramoto, KuramotoSakaguchi  # noqa: F401
-from .lotka_voltera import LotkaVoltera, LotkaVolteraSDE  # noqa: F401
-from .planted_aquarium import PlantedTankNitrogenCycle  # noqa: F401
+)
+from .geometric_brownian_motion import GeometricBrownianMotion
+from .kuramoto import Kuramoto, KuramotoSakaguchi
+from .lotka_voltera import LotkaVoltera, LotkaVolteraSDE
+from .planted_aquarium import PlantedTankNitrogenCycle
 
-from .michaelis_menten import MichaelisMenten  # noqa: F401
-from .mutualistic_pop import MutualisticPopulation  # noqa: F401
-from .ornstein_uhlenbeck import OrnsteinUhlenbeck  # noqa: F401
+from .michaelis_menten import MichaelisMenten
+from .mutualistic_pop import MutualisticPopulation
+from .ornstein_uhlenbeck import OrnsteinUhlenbeck
 from .quadratic_sdes import (
     Belozyorov3DQuad,
     Liping3DQuadFinance,
@@ -35,14 +38,14 @@ from .quadratic_sdes import (
     Rossler,
     Thomas,
     MooreSpiegel,
-)  # noqa: F401
+)
 from .simple_linear_sdes import (
     DampedOscillator,
     LinearSDE,
     imag_roots_2d_linear_sde,
     imag_roots_4d_linear_sde,
     attracting_fixed_point_4d_linear_sde,
-)  # noqa: F401
-from .sis import SIS  # noqa: F401
-from .statespace_models import VARMADynamics  # noqa: F401
-from .wilson_cowan import WilsonCowan  # noqa: F401
+)
+from .sis import SIS
+from .statespace_models import VARMADynamics
+from .wilson_cowan import WilsonCowan
