+++ interfere/__init__.py
@@ -1,16 +1,19 @@
-from .base import DynamicModel, ForecastMethod  # noqa: F401
-from .cross_validation import CrossValObjective  # noqa: F401
-from . import dynamics  # noqa: F401
+"""Interfere package."""
+# flake8: noqa
+
+from .base import DynamicModel, ForecastMethod
+from .cross_validation import CrossValObjective
+from . import dynamics
 from .interventions import (
-    perfect_intervention,  # noqa: F401
-    signal_intervention,  # noqa: F401
-    PerfectIntervention,  # noqa: F401
-    SignalIntervention,  # noqa: F401
-    ExogIntervention,  # noqa: F401
-    IdentityIntervention,  # noqa: F401
+    perfect_intervention,
+    signal_intervention,
+    PerfectIntervention,
+    SignalIntervention,
+    ExogIntervention,
+    IdentityIntervention,
 )
-from . import metrics  # noqa: F401
-from ._methods.sindy import SINDy  # noqa: F401
-from ._methods.vector_autoregression import VAR  # noqa: F401
-from ._methods.reservoir_computer import ResComp  # noqa: F401
-from . import utils  # noqa: F401
+from . import metrics
+from ._methods.sindy import SINDy
+from ._methods.vector_autoregression import VAR
+from ._methods.reservoir_computer import ResComp
+from . import utils
