+++ docs/index.md
@@ -1,16 +1,26 @@
 # 🌀 Interfere: Intervention Response Prediction in Complex Dynamic Models
-[![PyPI Version](https://img.shields.io/pypi/v/interfere)](https://pypi.org/project/interfere/) [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](LICENSE)
+
+[![PyPI Version](https://img.shields.io/pypi/v/interfere)](https://pypi.org/project/interfere/) [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT) [![Documentation](https://img.shields.io/badge/docs-gh--pages-blue)](https://djpasseyjr.github.io/interfere/)
 
 Interfere is a comprehensive Python toolkit for simulating, intervening on, and
-optimizing forecasting methods to predict the behavior complex dynamical systems. It enables:
+optimizing forecasting methods to predict the behavior of complex dynamical systems. It enables:
 
 - Rich collections of continuous and discrete-time dynamic models (ODEs, SDEs,
-  differnce equations, and more).
+  difference equations, and more).
 - Exogenous interventions to generate control and treatment scenarios at scale.
 - Seamless forecasting integrations (SINDy, VAR, reservoir computing, ARIMA, LSTM, NHITS).
 - Automated sliding-window cross-validation and hyperparameter tuning with Optuna.
 - Unified error metrics and evaluation workflows for intervention-response
   prediction.
+
+## Overview
+
+Interfere is a research-oriented Python package that addresses a fundamental question in complex systems: *When can we predict how complex systems will respond to interventions?* This package provides tools for:
+
+- Modeling dynamic nonlinear multivariate stochastic systems.
+- Simulating and analyzing how such systems respond to interventions.
+- Generating complex dynamic counterfactuals.
+- Studying causal relationships in complex systems.
 
 ## Interfere Benchmark Dataset ([Download](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing))
 
@@ -23,9 +33,13 @@ available for download as the [Interfere Benchmark
 It can be used to benchmark a forecasting method's ability to predict the
 response of a dynamic system to interventions.
 
+## Documentation
+
+Full documentation is built with MkDocs and published at **[https://djpasseyjr.github.io/interfere/](https://djpasseyjr.github.io/interfere/)**.
+
 ## Installation
 
-Install the core package from PyPI:
+### From PyPI
 
 ```bash
 pip install interfere
@@ -35,6 +49,14 @@ For Nixtla-based forecasting methods (`ARIMA`, `LSTM`, `NHITS`), install the ext
 
 ```bash
 pip install interfere[nixtla]
+```
+
+### From Local Clone
+
+```bash
+git clone https://github.com/djpasseyjr/interfere.git
+cd interfere
+pip install .
 ```
 
 ## Quick Start
@@ -139,9 +161,139 @@ pred_traj = method.simulate(
 
 The SINDy method identifies the underlying dynamics of the system using sparse regression techniques, making it particularly effective for discovering interpretable mathematical models of complex systems.
 
-## Documentation
+## Dependencies
+
+1. Basic: `pip install interfere`
+2. Full forecasting methods: `pip install "interfere[methods]"`
+3. Developer / experimental features: `pip install "interfere[dev]"`
+
+## Example Use
+
+The package can be used to simulate and analyze how systems respond to interventions. For example, it can model the effect of stochasticity on intervention response forecasting:
+
[…]
