+++ .github/workflows/python-package.yml
@@ -40,3 +40,6 @@ jobs:
       uses: codecov/codecov-action@v5
       with:
           token: ${{ secrets.CODECOV_TOKEN }}
+          
+    - name: Build docs
+      run: mkdocs build --strict

[…]
