### 2026-02-09 5edad8b
Adds benchmark to docs
