+++ docs/benchmark.md
@@ -1,0 +1,137 @@
+# Working with the benchmark
+
+The [Interfere Benchmark](https://drive.google.com/file/d/19_Ha-D8Kb1fFJ_iECU62eawbeuCpeV_g/view?usp=sharing) is a set of JSON files. Each file encodes one **observational vs intervention** problem: you get training data, an **observational** trajectory (continuation with the same exogenous driver), and an **intervention** trajectory (the same system under an intervention). The goal is to fit a forecasting method on the training data and compare how well it predicts the observational continuation versus the intervention response.
+
+## What's in the JSON
+
+Each benchmark JSON includes (among other keys) the following at the top level:
+
+| Key                          | Description                                                                     |
+| ---------------------------- | ------------------------------------------------------------------------------- |
+| `metadata`                 | Variable descriptions and schema info                                           |
+| `model_description`        | Text description of the underlying dynamics                                     |
+| `initial_condition_times`  | Times for the short history before training                                     |
+| `initial_condition_states` | States at those times (prior to `train_*`)                                    |
+| `train_times`              | 1D array of training time points                                                |
+| `train_states`             | 2D array `(n_train, n_vars)` of training observations                         |
+| `train_exog_idxs`          | Column indices of exogenous variables during training (same as forecast)        |
+| `forecast_times`           | Time points for the forecast period                                             |
+| `forecast_states`          | **Observational** trajectory: system continued with same exogenous driver |
+| `forecast_exog_idxs`       | Same as `train_exog_idxs`                                                     |
+| `causal_resp_states`       | **Intervention** trajectory: response under the "do" intervention         |
+| `causal_resp_times`        | Same length as `forecast_times` (same time grid)                              |
+| `causal_resp_exog_idxs`    | Exogenous indices under the intervention (includes any newly driven variables)  |
+| `target_idx`               | (Optional) column index of the prediction target                                |
+
+The **intervention is not stored as a separate object**. It is implied by the exogenous columns: during the observational period they follow `train_states` / `forecast_states` at `forecast_exog_idxs`; during the intervention period they follow `train_states` / `causal_resp_states` at `causal_resp_exog_idxs`. You reconstruct callable interventions by interpolating those time series (see below).
+
+## The three stages
+
+1. **Training**Fit a forecasting method on `train_states` at `train_times`, with exogenous variables at `train_exog_idxs`.
+2. **Observational forecast**Using the **same** exogenous driver as in training (values from the observational run: training then `forecast_states`), simulate over `forecast_times` and compare to `forecast_states`.
+3. **Intervention forecast**
+   Using the **intervention** exogenous driver (values from the intervention run: training then `causal_resp_states` at `causal_resp_exog_idxs`), simulate over the same `forecast_times` and compare to `causal_resp_states`.
+
+So: same training data; one continuation is "observational", the other is "under intervention"; the benchmark compares how well a method predicts both.
+
+## Loading the JSON and building interventions
+
+The package expects interventions as **callables** of time (e.g. `SignalIntervention` with scalar functions). The JSON only gives (time, value) arrays. So we load the arrays and turn the exogenous columns into callables via interpolation (e.g. `scipy.interpolate.interp1d`):
+
+```python
+import json
[…]
