### 2026-02-09 56c0089
Adds setuptools for pysindy on python >= 3.11
