### 2026-02-13 #9792 @mstimberg
@Jacob-Stevens-Haas Apologies for the late reply… You are not off base on anything, thank you very much for the thorough review of the paper, and I've seen that you also jumped into the PR to comment on issues of the code and packaging. From my side, the review process is looking perfectly on tracks 👍
