### 2026-02-25 10993a9
Enhance hyperparameter optimization explanation
Clarified the description of hyperparameter optimization in forecasting methods and added context regarding observational data and interventions.
