+++ paper/paper_full.md
@@ -141,9 +141,11 @@ the natural evolution behavior of the system. \label{fig:interv_effect}](../imag
 Interfere offers tools to optimize forecasting methods for time series
 prediction. By using Interfere's cross validation objective function along with a
 hyperparameter optimizer such as Optuna [@akiba2019optuna], it is possible to compare
-hyperparameter settings on multiple folds of time series data. To simplify this
-process, every Interfere forecasting method comes with sensible preset
-hyperparameter ranges for the optimizer to explore. Using the cross validation objective function for hyperparameter optimization is demonstrated in the following code block.
+hyperparameter settings on multiple folds of time series data, the data can be purely observational 
+(as in the benchmark) or contain interventions and responses. 
+Every Interfere forecasting method comes with sensible preset
+hyperparameter ranges for the optimizer to explore. Hyperparameter optimization of cross validated error
+is demonstrated in the following code block.
 
 ```python
 
