### 2026-02-25 6b7b8bb
Update figure caption for clarity
Clarified figure caption regarding the simulation of the quadratic Belozyorov system and the representation of stochastic noise.
