### 2026-02-25 305c1d5
Add new references to paper.bib
