+++ paper/paper.bib
@@ -522,3 +522,30 @@ journal = {Journal of Open Source Software}
     note = {Publisher: Society for Industrial and Applied Mathematics},
     pages = {922--952},
 }
+
+@article{kaptanoglu2023benchmarking,
+  author  = {Kaptanoglu, Alan A. and Zhang, Linan and Nicolaou, Zachary G. and Fasel, Urban and Brunton, Steven L.},
+  title   = {Benchmarking sparse system identification with low-dimensional chaos},
+  journal = {Nonlinear Dynamics},
+  volume  = {111},
+  number  = {14},
+  pages   = {13143--13164},
+  year    = {2023},
+  month   = jul,
+  doi     = {10.1007/s11071-023-08525-4}
+}
+
+@article{PhysRevResearch.5.043252,
+  title = {Model scale versus domain knowledge in statistical forecasting of chaotic systems},
+  author = {Gilpin, William},
+  journal = {Phys. Rev. Res.},
+  volume = {5},
+  issue = {4},
+  pages = {043252},
+  numpages = {18},
+  year = {2023},
+  month = {Dec},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevResearch.5.043252},
+  url = {https://link.aps.org/doi/10.1103/PhysRevResearch.5.043252}
+}
