### 2026-02-25 4f2c1dc
Update BibTeX entries for articles
