+++ paper/paper.bib
@@ -523,7 +523,7 @@ journal = {Journal of Open Source Software}
     pages = {922--952},
 }
 
-@article{kaptanoglu2023benchmarking,
+@article{kaptanoglu_benchmarking_2023,
   author  = {Kaptanoglu, Alan A. and Zhang, Linan and Nicolaou, Zachary G. and Fasel, Urban and Brunton, Steven L.},
   title   = {Benchmarking sparse system identification with low-dimensional chaos},
   journal = {Nonlinear Dynamics},
@@ -535,7 +535,7 @@ journal = {Journal of Open Source Software}
   doi     = {10.1007/s11071-023-08525-4}
 }
 
-@article{PhysRevResearch.5.043252,
+@article{gilpin_model_2023,
   title = {Model scale versus domain knowledge in statistical forecasting of chaotic systems},
   author = {Gilpin, William},
   journal = {Phys. Rev. Res.},
