### 2026-02-25 5fa2a8d
Added Gilpin and Kaptanoglu citations.
