+++ paper/paper.md
@@ -64,39 +64,66 @@ These sixty scenarios comprise the downloadable [Interfere Benchmark 1.1.1](http
 **System Trajectory After Intervention (Center):** Exogenous control of $x(t)$ by $\text{do}(x(t)=\sin(t))$ for $t>10$ affects $y(t)$. 
 The intervention (black) and response (blue) indicate clear departure from natural behavior (green and gray). 
 **Intervention Response Prediction (Right):**
-Equation discovery by SINDy [@brunton_discovering_2016] is fit to data prior to the intervention to predict intervention response (red).](../images/interfere_usage_combined.png)
+Equation discovery by SINDy [@brunton_discovering_2016] is fit to data prior to
+the intervention to predict intervention response (red).
+\label{fig:interv_example}](../images/interfere_usage_combined.png)
 
 Multiple frameworks have emerged for identifying causal relationships from data [@imbens_causal_2015; @pearl_causality_2009; @wieczorek_information_2019].
 Many frameworks assume a static, linear relationship.
 Because a linear approximation can be insufficient for anticipating the response of complex dynamic systems at medium and long times, 
 non-linear, dynamic techniques have been developed for causal discovery and forecasting [e.g. @runge_discovering_2022]. Nevertheless, there are relatively few intervention oriented techniques to fit dynamic nonlinear models to data. 
 
-Interfere's benchmarking approach is enabled by a unified toolkit for simulating intervention response, standardized method evaluation, and systematic cross-validation across diverse dynamic systems. 
+Interfere's benchmarking approach is enabled by a unified toolkit for simulating
+intervention response, standardized method evaluation, and systematic
+cross-validation across diverse dynamic systems (see Figure
+\ref{fig:interv_example} for an example scenario). 
 By lowering barriers to method evaluation, Interfere facilitates  development of prediction methods that translate from theory to real-world application.
 
 # Primary Contributions
 
 Interfere provides (1) dynamically diverse counterfactuals, (2) cross-disciplinary forecast methods, and (3) comprehensive and extensible benchmarking.
 
-![Can stochasticity help reveal associations between variables? In this example, Interfere compares intervention response prediction for deterministic and stochastic versions of the same system.
+![Can stochasticity help reveal associations between variables? This example
+captures the kind of hypothesis that can be tested using Interfere's ability to compare intervention response prediction for deterministic
+and stochastic versions of the same system.
  \label{fig:det_vs_stoch}](../images/det_v_stoch.png)
 
 ## 1. Dynamically Diverse Counterfactuals at Scale
 
-Whereas many predictive methods are typically benchmarked on $<10$ systems [@challu_nhits_2023; @brunton_discovering_2016; @vlachas_backpropagation_2020; @pathak_model-free_2018; @prasse_predicting_2022], Interfere's "dynamics" submodule contains $>50$ models, including linear, nonlinear, chaotic, continuous-time, discrete-time, stochastic, and deterministic models, from disciplines including finance, ecology, biology, neuroscience and public health. Most importantly, Interfere is built for studying interventions: the inherited "BaseDynamics" type allows exogenous control of any observed state, added measurement noise, and, for most models, stochasticity parameterized by a scalar or covariance matrix. Interfere thus offers a user-friendly framework to produce complex dynamic intervention response and forecasting scenarios at scale.
+Whereas many predictive methods are typically benchmarked on $<10$ systems
+[@challu_nhits_2023; @brunton_discovering_2016; @vlachas_backpropagation_2020;
+@pathak_model-free_2018; @prasse_predicting_2022], Interfere's "dynamics"
+submodule contains $>50$ models, including linear, nonlinear, chaotic,
+continuous-time, discrete-time, stochastic, and deterministic models, from
[…]
