+++ paper/paper.bib
@@ -23,7 +23,7 @@
 	doi = {10.12732/ijpam.v114i1.12},
 	file = {Banks et al. - 2017 - PARAMETER ESTIMATION FOR AN ALLOMETRIC FOOD WEB MO.pdf:/Users/djpassey/Zotero/storage/3X6FKEMC/Banks et al. - 2017 - PARAMETER ESTIMATION FOR AN ALLOMETRIC FOOD WEB MO.pdf:application/pdf},
 	issn = {1311-8080, 1314-3395},
-	journal = {International Journal of Pure and Apllied Mathematics},
+	journal = {International Journal of Pure and Applied Mathematics},
 	language = {en},
 	month = apr,
 	number = {1},
@@ -217,7 +217,7 @@
 }
 
 @article{hening_stochastic_2018,
-	title = {Stochastic lotka–volterra food chains},
+	title = {Stochastic Lotka–Volterra food chains},
 	volume = {77},
 	journal = {Journal of mathematical biology},
 	author = {Hening, Alexandru and Nguyen, Dang H},
