### 2026-03-18 #9792 @Jacob-Stevens-Haas
Thanks for making the changes!

So I apparently reviewed the longer-form paper, but @djpasseyjr adapted the comments to the version for submission.  Everything looks good on the paper front, with the caveat that the package includes a `ForecastMethod` class but paper refers to  a `ForecastingMethod` class.

IDK if the gh-pages documentation is in-scope, but it looks like it needs to be rebuilt to include the new files, custom_forecast_methods.md and benchmark.md.  The issue I added for documentation also may have been closed prematurely; [this comment](https://github.com/djpasseyjr/interfere/issues/8#issuecomment-3831877542) and forwards may not have been seen, and remain in the current version.
