### 2026-03-30 f64648d
Fixed minor errors in readme.
