### 2026-03-30 bd88924
Normalize Lotka-Volterra spelling and naming
Apply spelling renames from Voltera to Volterra across code/docs, including module rename and updated references.
