+++ docs/index.md
@@ -150,7 +150,7 @@ Y_endog, Y_exog = intervention.split_exog(sim_states)
 # Fit SINDy to the pre-intervention data.
 method.fit(t_train, Y_endog, Y_exog)
 
-# Use the inherited interfere.ForecastingMethod.simulate() method
+# Use the inherited interfere.ForecastMethod.simulate() method
 # To simulate intervention response using SINDy
 pred_traj = method.simulate(
     test_t, prior_states=sim_states, intervention=intervention
@@ -270,7 +270,7 @@ python -m pytest -v tests
 
 ## License
 
-This project is licensed under the MIT License - see the [LICENSE](LICENSE) file for details.
+This project is licensed under the MIT License - see the [LICENSE](../LICENSE) file for details.
 
 ## Citation
 
