### 2026-03-30 #9792 @djpasseyjr
Thanks @Jacob-Stevens-Haas for flagging the issues that slipped through. I've corrected them [here](https://github.com/djpasseyjr/interfere/pull/11).

Anything to add @yunus-kausable ?
