### 2026-04-09 #9792 @yunus-kausable
Hi @djpasseyjr,

thanks for the work on this package and for submitting it to JOSS! I've completed my review and ticked off the remaining items on my checklist. I was looking at the `JOSS-Review` branch (currently 9 commits ahead of `main`), which would still need to be merged once the review is concluded.

I think `interfere` fills a really interesting gap by allowing to systematically benchmark how well forecasting methods may predict intervention response... and the variety of dynamical systems to test this on is really cool! The package has a clean, unified interface for simulation, intervention and prediction, and is well-tested. And the software paper is a good starting point to get started with the package.
I'm happy to **recommend acceptance** of this submission.
