### 2026-04-09 #9792 @mstimberg
@yunus-kausable, many thanks for your review, I've noted your recommendation.

@Jacob-Stevens-Haas how are things looking from your side, are you happy with the way https://github.com/djpasseyjr/interfere/pull/11 addresses your comments?
