### 2026-04-28 #9792 @danielskatz
👋 @mstimberg  - As track chair, I try to check on submissions in the CSISM track semi-regularly. For this one, what's its status? I can't see that anything has happened in the 2-3 weeks.
