### 2026-04-29 #9792 @mstimberg
@danielskatz Yes, apologies, I haven't been following up this review as closely as I should have.

@Jacob-Stevens-Haas Could you please get back to us whether you are happy with the way djpasseyjr/interfere#11 addresses your comments and whether there is anything else that you think should be improved before recommending the software/paper for publication?
