### 2026-05-21 4bfaab1
Release 1.0.2 for JOSS review.
