+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "interfere"
-version = "1.0.1"
+version = "1.0.2"
 description = "Predicting how complex systems will respond to interventions."
 readme = "README.md"
 license = { file = "LICENSE" }
