### 2026-05-21 919eef0
Maintain PyPI compatibility with deps.
