### 2026-06-10 4f64966
Update README links and docs build notes
