### 2026-06-10 0bdb93f
Move extended paper into MkDocs docs
