+++ docs/extended_paper.bib
@@ -1,0 +1,551 @@
+@article{baker_mechanistic_2018,
+    title = {Mechanistic models versus machine learning, a fight worth fighting for the biological community?},
+    volume = {14},
+    issn = {1744-9561, 1744-957X},
+    url = {https://royalsocietypublishing.org/doi/10.1098/rsbl.2017.0660},
+    doi = {10.1098/rsbl.2017.0660},
+    abstract = {Ninety per cent of the world's data have been generated in the last 5 years (
+              Machine learning: the power and promise of computers that learn by example
+              . Report no. DES4702. Issued April 2017. Royal Society). A small fraction of these data is collected with the aim of validating specific hypotheses. These studies are led by the development of mechanistic models focused on the causality of input–output relationships. However, the vast majority is aimed at supporting statistical or correlation studies that bypass the need for causality and focus exclusively on prediction. Along these lines, there has been a vast increase in the use of machine learning models, in particular in the biomedical and clinical sciences, to try and keep pace with the rate of data generation. Recent successes now beg the question of whether mechanistic models are still relevant in this area. Said otherwise, why should we try to understand the mechanisms of disease progression when we can use machine learning tools to directly predict disease outcome?},
+    language = {en},
+    number = {5},
+    urldate = {2025-03-19},
+    journal = {Biology Letters},
+    author = {Baker, Ruth E. and Peña, Jose-Maria and Jayamohan, Jayaratnam and Jérusalem, Antoine},
+    month = may,
+    year = {2018},
+    pages = {20170660},
+}
+
+@article{banks_parameter_2017,
+	abstract = {The application of mechanistic models to natural systems is of interest to ecological researchers. We use the mechanistic Allometric Trophic Network (ATN) model, which is well-studied for controlled and theoretical systems, to describe the dynamics of the aphid Rhopalosiphum padi in an agricultural field. We diagnose problems that arise in a first attempt at a least squares parameter estimation on this system, including formulation of the model for the inverse problem and information content present in the data. We seek to establish whether the field data, as it is currently collected, can support parameter estimation for the ATN model.},
+	author = {Banks, H.T. and Banks, J.E. and Bommarco, R. and Curtsdotter, A. and Jonsson, T. and Laubmeier, A.N.},
+	doi = {10.12732/ijpam.v114i1.12},
+	file = {Banks et al. - 2017 - PARAMETER ESTIMATION FOR AN ALLOMETRIC FOOD WEB MO.pdf:/Users/djpassey/Zotero/storage/3X6FKEMC/Banks et al. - 2017 - PARAMETER ESTIMATION FOR AN ALLOMETRIC FOOD WEB MO.pdf:application/pdf},
+	issn = {1311-8080, 1314-3395},
+	journal = {International Journal of Pure and Applied Mathematics},
+	language = {en},
+	month = apr,
+	number = {1},
+	title = {Parameter estimation for an allometric food web model},
+	url = {http://www.ijpam.eu/contents/2017-114-1/12/},
+	urldate = {2023-10-06},
+	volume = {114},
+	year = {2017},
+	bdsk-url-1 = {http://www.ijpam.eu/contents/2017-114-1/12/},
+	bdsk-url-2 = {https://doi.org/10.12732/ijpam.v114i1.12}
+}
+
+@article{brayton_frbus_2014,
+    title = {The {FRB}/{US} model: A tool for macroeconomic policy analysis},
+    number = {2014-04},
+    journal = {FEDS Notes},
+    author = {Brayton, Flint and Laubach, Thomas and Reifschneider, David},
+    year = {2014},
+    pages = {03},
+	doi={10.17016/2380-7172.0012},
+}
+
+@book{imbens_causal_2015,
+	edition = {1},
[…]
