### 2026-06-10 1891773
Render extended paper citations via mkdocs-bibtex
