+++ .github/workflows/python-package.yml
@@ -32,6 +32,11 @@ jobs:
         pip install ".[dev]"
         pip install pytest-cov
 
+    - name: Install git-only test dependencies
+      run: |
+        pip install git+https://github.com/djpasseyjr/surd.git
+        pip install pyclustering@git+https://github.com/djpasseyjr/pyclustering
+
     - name: Test with pytest
       run: |
         pytest --cov --cov-branch --cov-report=xml tests
@@ -41,5 +46,8 @@ jobs:
       with:
           token: ${{ secrets.CODECOV_TOKEN }}
           
+    - name: Install pandoc (required by mkdocs-bibtex)
+      run: sudo apt-get update && sudo apt-get install -y pandoc
+
     - name: Build docs
       run: mkdocs build --strict
