### 2026-06-10 7276c8f
Trim JOSS paper.bib and link extended version in docs
