+++ paper/paper.bib
@@ -137,22 +137,6 @@
     pages = {975},
 }
 
-
-@article{cliff_unifying_2023,
-	title = {Unifying pairwise interactions in complex dynamics},
-	volume = {3},
-	issn = {2662-8457},
-	url = {https://doi.org/10.1038/s43588-023-00519-x},
-	doi = {10.1038/s43588-023-00519-x},
-	abstract = {Scientists have developed hundreds of techniques to measure the interactions between pairs of processes in complex systems, but these computational methods—from contemporaneous correlation coefficients to causal inference methods—define and formulate interactions differently, using distinct quantitative theories that remain largely disconnected. Here we introduce a large assembled library of 237 statistics of pairwise interactions, and assess their behavior on 1,053 multivariate time series from a wide range of real-world and model-generated systems. Our analysis highlights commonalities between disparate mathematical formulations of interactions, providing a unified picture of a rich interdisciplinary literature. Using three real-world case studies, we then show that simultaneously leveraging diverse methods can uncover those most suitable for addressing a given problem, facilitating interpretable understanding of the quantitative formulation of pairwise dependencies that drive successful performance. Our results and accompanying software enable comprehensive analysis of time-series interactions by drawing on decades of diverse methodological contributions.},
-	number = {10},
-	journal = {Nature Computational Science},
-	author = {Cliff, Oliver M. and Bryant, Annie G. and Lizier, Joseph T. and Tsuchiya, Naotsugu and Fulcher, Ben D.},
-	month = oct,
-	year = {2023},
-	pages = {883--893},
-}
-
 @article{prasse_predicting_2022,
 	title = {Predicting network dynamics without requiring the knowledge of the interaction graph},
 	volume = {119},
@@ -170,102 +154,6 @@
 	file = {Full Text PDF:/Users/djpassey/Zotero/storage/NRWUWKL2/Prasse and Van Mieghem - 2022 - Predicting network dynamics without requiring the .pdf:application/pdf},
 }
 
-@book{oksendal_stochastic_2005,
-	address = {Berlin},
-	edition = {6th ed., corrected third printing 2005},
-	series = {Universitext},
-	title = {Stochastic differential equations: an introduction with applications},
-	isbn = {978-3-540-04758-2},
-	shorttitle = {Stochastic differential equations},
-	language = {eng},
-	publisher = {Springer},
-	author = {Øksendal, Bernt Karsten},
-	year = {2005},
-}
-
-@article{lloyd_coupled_1995,
-	title = {The coupled logistic map: a simple model for the effects of spatial heterogeneity on population dynamics},
-	volume = {173},
-	number = {3},
-	journal = {Journal of Theoretical Biology},
-	author = {Lloyd, Alun L},
-	year = {1995},
-	note = {Publisher: Elsevier},
-	pages = {217--230},
-	doi = {10.1006/jtbi.1995.0058},
-}
-
-@incollection{kaneko_coupled_1991,
-	title = {Coupled map lattice},
-	booktitle = {Chaos, Order, and Patterns},
-	publisher = {Springer},
-	author = {Kaneko, Kunihiko},
-	year = {1991},
-	pages = {237--247},
-}
-
-@article{srinivasan_guide_2022,
-	title = {A guide to the {Michaelis}–{Menten} equation: steady state and beyond},
-	volume = {289},
-	number = {20},
-	journal = {The FEBS journal},
-	author = {Srinivasan, Bharath},
-	year = {2022},
[…]
