### 2026-06-11 be33f77
Adds {P} for python in the bibliography.
