+++ paper/paper.bib
@@ -195,14 +195,14 @@ volume = {7},
 number = {69},
 pages = {3994},
 author = {Alan A. Kaptanoglu and Brian M. de Silva and Urban Fasel and Kadierdan Kaheman and Andy J. Goldschmidt and Jared Callaham and Charles B. Delahunt and Zachary G. Nicolaou and Kathleen Champion and Jean-Christophe Loiseau and J. Nathan Kutz and Steven L. Brunton},
-title = {PySINDy: A comprehensive Python package for robust sparse system identification},
+title = {PySINDy: A comprehensive {P}ython package for robust sparse system identification},
 journal = {Journal of Open Source Software}
 }
 
 @inproceedings{seabold2010statsmodels,
-  title={statsmodels: Econometric and statistical modeling with python},
+  title={statsmodels: Econometric and statistical modeling with {P}ython},
   author={Seabold, Skipper and Perktold, Josef},
-  booktitle={9th Python in Science Conference},
+  booktitle={9th {P}ython in Science Conference},
   year={2010},
   doi = {10.25080/majora-92bf1922-011},
 }
