### 2026-06-16 #9792 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
